Remove all old code (#10)
