+++ Sources/MQTTNIO/ChannelHandlers/MQTTMessageDecoder.swift
@@ -22,20 +22,10 @@ struct ByteToMQTTMessageDecoder: NIOSingleStepByteToMessageDecoder {
 
     typealias InboundOut = MQTTPacket
 
-    let version: MQTTClient.Version
+    let version: MQTTConnectionConfiguration.Version
 
-    init(version: MQTTClient.Version) {
-        self.version = version
-    }
-
-    // TODO: remove once MQTTClient.Version is substituted everywhere
     init(version: MQTTConnectionConfiguration.Version) {
-        switch version {
-        case .v3_1_1:
-            self.version = .v3_1_1
-        case .v5_0:
-            self.version = .v5_0
-        }
+        self.version = version
     }
 
     mutating func decode(buffer: inout ByteBuffer) throws -> MQTTPacket? {
